# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a API for ASR DEMO.",
        "title": "ASR DEMO API",
        "contact": {
            "name": "API Support",
            "email": "wayne900619@gmail.com"
        },
        "version": "0.1"
    },
    "host": "asrvm.iis.sinica.edu.tw:8080",
    "basePath": "/api",
    "paths": {
        "/postRecognize": {
            "post": {
                "description": "get post recognize result",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Do post recognize",
                "parameters": [
                    {
                        "description": "Set AsrKind to model name you want to use in post recognize",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.segmentInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.wordalignment"
                            }
                        }
                    },
                    "400": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
        "/translate": {
            "post": {
                "description": "get translate result",
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "text/plain"
                ],
                "summary": "Do translate after post recognize",
                "parameters": [
                    {
                        "description": "string that need to translate",
                        "name": "string",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "return translated string",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
        "/uploadRecognize": {
            "post": {
                "description": "get upload recognize result",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Do upload recognize",
                "parameters": [
                    {
                        "type": "string",
                        "name": "asrKind",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "format": "binary",
                        "name": "file",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "langKind",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.wordalignment"
                            }
                        }
                    },
                    "400": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.segmentInfo": {
            "type": "object",
            "properties": {
                "asrKind": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "langKind": {
                    "type": "string"
                },
                "length": {
                    "type": "number"
                },
                "start": {
                    "type": "number"
                }
            }
        },
        "handler.wordalignment": {
            "type": "object",
            "properties": {
                "length": {
                    "type": "number"
                },
                "start": {
                    "type": "number"
                },
                "token": {
                    "type": "string"
                },
                "word": {
                    "type": "string"
                }
            }
        }
    }
}